requirements.txt
Flask==2.2.2
Werkzeug==2.3.7
psycopg2-binary==2.9.6